requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
atomicwrites==1.3.0
attrs==19.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
docutils==0.14
idna==2.8
importlib-metadata==0.18
more-itertools==7.1.0
packaging==19.0
pinyin==0.4.0
pkginfo==1.5.0.1
pluggy==0.12.0
py==1.8.0
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==5.0.1
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
tqdm==4.32.2
twine==1.13.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
xlrd==1.2.0
zipp==0.5.2
